apply MVC on backend code
